+++ plugins/zentao/models/archived/project.go
@@ -82,9 +82,9 @@ type ZentaoProject struct {
 	TeamCount      int    `json:"teamCount"`
 	LeftTasks      string `json:"leftTasks"`
 	//TeamMembers   []interface{} `json:"teamMembers" gorm:"-"`
-	TotalEstimate int     `json:"totalEstimate"`
-	TotalConsumed int     `json:"totalConsumed"`
-	TotalLeft     int     `json:"totalLeft"`
+	TotalEstimate float64 `json:"totalEstimate"`
+	TotalConsumed float64 `json:"totalConsumed"`
+	TotalLeft     float64 `json:"totalLeft"`
 	Progress      float64 `json:"progress"`
 	TotalReal     int     `json:"totalReal"`
 }
@@ -101,9 +101,9 @@ type Whitelist []struct {
 	WhitelistRealname string `json:"realname"`
 }
 type Hours struct {
-	HoursTotalEstimate int     `json:"totalEstimate"`
-	HoursTotalConsumed int     `json:"totalConsumed"`
-	HoursTotalLeft     int     `json:"totalLeft"`
+	HoursTotalEstimate float64 `json:"totalEstimate"`
+	HoursTotalConsumed float64 `json:"totalConsumed"`
+	HoursTotalLeft     float64 `json:"totalLeft"`
 	HoursProgress      float64 `json:"progress"`
 	HoursTotalReal     int     `json:"totalReal"`
 }
